arm/gic: refactor types
the main goal is to make the gic drivers as portable as possible.
Fixed-width types were matched to the correct width in the gic
specfication and arch-agnostic types were used for registers which the
spec matches to machine width

+++ src/arch/armv8/gic.c
@@ -87,7 +87,7 @@ void gic_init()
 
 void gic_handle()
 {
-    uint64_t ack = gicc_iar();
+    uint32_t ack = gicc_iar();
     irqid_t id = bit32_extract(ack, GICC_IAR_ID_OFF, GICC_IAR_ID_LEN);
 
     if (id < GIC_FIRST_SPECIAL_INTID) {
@@ -97,12 +97,12 @@ void gic_handle()
     }
 }
 
-uint64_t gicd_get_prio(irqid_t int_id)
+uint8_t gicd_get_prio(irqid_t int_id)
 {
     size_t reg_ind = GIC_PRIO_REG(int_id);
     size_t off = GIC_PRIO_OFF(int_id);
 
-    uint64_t prio =
+    uint8_t prio =
         gicd.IPRIORITYR[reg_ind] >> off & BIT32_MASK(off, GIC_PRIO_BITS);
 
     return prio;
@@ -112,7 +112,7 @@ void gicd_set_icfgr(irqid_t int_id, uint8_t cfg)
 {
     size_t reg_ind = (int_id * GIC_CONFIG_BITS) / (sizeof(uint32_t) * 8);
     size_t off = (int_id * GIC_CONFIG_BITS) % (sizeof(uint32_t) * 8);
-    uint64_t mask = ((1U << GIC_CONFIG_BITS) - 1) << off;
+    uint32_t mask = ((1U << GIC_CONFIG_BITS) - 1) << off;
 
     spin_lock(&gicd_lock);
 
@@ -125,7 +125,7 @@ void gicd_set_prio(irqid_t int_id, uint8_t prio)
 {
     size_t reg_ind = GIC_PRIO_REG(int_id);
     size_t off = GIC_PRIO_OFF(int_id);
-    uint64_t mask = BIT32_MASK(off, GIC_PRIO_BITS);
+    uint32_t mask = BIT32_MASK(off, GIC_PRIO_BITS);
 
     spin_lock(&gicd_lock);
 
@@ -169,7 +169,7 @@ bool gicd_get_act(irqid_t int_id)
 void gicd_set_enable(irqid_t int_id, bool en)
 {
     size_t reg_ind = GIC_INT_REG(int_id);
-    uint64_t bit = GIC_INT_MASK(int_id);
+    uint32_t bit = GIC_INT_MASK(int_id);
 
     if (en) {
         gicd.ISENABLER[reg_ind] = bit;
